Products_RAG_main/requirements.txt
# certifi==2025.10.5
chromadb==1.3.4
# numpy==2.3.4
openai==2.7.1
pandas==2.3.3
protobuf==6.33.0
pymongo==4.15.3
python-dotenv==1.2.1
qdrant_client==1.15.1
Requests==2.32.5
sentence_transformers==5.1.1
supabase==2.24.0
# torch==2.8.0
